+++ RELEASE_NOTES.md
@@ -3,8 +3,8 @@ Release notes
 
 Changes in version 0.4.8
 ========================
-  - Changes to departments API: SearchReply::register_departments() method now takes parent and current department argument,
-    and uses Department::SCPtr for both. SearchListenerBase::push() method for departments got changed to match as well. Removed
+  - Changes to departments API: SearchReply::register_departments() method now takes parent department argument only,
+    and uses Department::SCPtr for it. SearchListenerBase::push() method for departments got changed to match as well. Removed
     constructors of Department and added static create() methods instead. Changed DepartmentList to hold Department
     pointers instead of values. Changed Department::set_has_subdepartments() method to take bool value (true by default).